requirements.txt
Flask==2.3.3
pytest==7.4.2
pytest-mock==3.11.1
pytest-cov==4.1.0
requests==2.31.0
pytest-playwright==0.4.3
